fix: build errors and compiler crash.
1. remove nullable primitives, assemblyscript does not support them.
2. remove nullable return from message (`owner_of`), compiler crash

+++ assembly/interfaces/balances.ts
@@ -1,9 +1,9 @@
-import { AccountId, Balance, SpreadLayout } from "ask-lang";
+import { AccountId, Balance, u128 } from "ask-lang";
 import { Id } from "../types";
 
 // export interface IBalances extends SpreadLayout{
 export interface IBalances {
-  balance_of(owner: AccountId): u32;
+  balance_of(owner: AccountId): u128;
   increase_balance(owner: AccountId, id: Id, increase_supply: bool): void;
   decrease_balance(owner: AccountId, id: Id, decrease_supply: bool): void;
   total_supply(): Balance;
